+++ sdk/communication/azure-communication-common/src/test/java/com/azure/communication/common/CommunicationTokenCredentialTests.java
@@ -3,11 +3,10 @@
 package com.azure.communication.common;
 
 import com.azure.communication.common.implementation.JwtTokenMocker;
-import com.azure.core.util.FluxUtil;
+import com.azure.communication.common.implementation.TokenParser;
+import com.azure.core.credential.AccessToken;
 import com.azure.core.util.logging.ClientLogger;
-
 import org.junit.jupiter.api.Test;
-
 import reactor.test.StepVerifier;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -16,6 +15,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.io.IOException;
+import java.time.Instant;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.function.Supplier;
@@ -58,6 +58,7 @@ public class CommunicationTokenCredentialTests {
 
     class MockImmediateRefresher implements Supplier<String> {
         private int numCalls = 0;
+        private String refreshedToken;
         private Runnable onCallReturn;
 
         public int numCalls() {
@@ -72,13 +73,19 @@ public class CommunicationTokenCredentialTests {
             numCalls = 0;
         }
 
+        public void setRefreshedToken(String refreshedToken) {
+            this.refreshedToken = refreshedToken;
+        }
+
         @Override
         public String get() {
             numCalls++;
             if (this.onCallReturn != null) {
                 this.onCallReturn.run();
             }
-            return tokenMocker.generateRawToken("Mock", "user", DEFAULT_EXPIRING_OFFSET_SECONDS);
+            return refreshedToken != null
+                    ? refreshedToken
+                    : tokenMocker.generateRawToken("Mock", "user", DEFAULT_EXPIRING_OFFSET_SECONDS);
         }
     }
 
@@ -94,11 +101,11 @@ public class CommunicationTokenCredentialTests {
                 .setInitialToken(tokenStr);
         CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
         StepVerifier.create(tokenCredential.getToken()).assertNext(token -> {
-                    assertFalse(token.isExpired(),
-                            "Refreshable AccessToken should not expire when expiry is set to 5 minutes later");
-                    assertEquals(tokenStr, token.getToken());
-                    assertEquals(0, immediateFresher.numCalls());
-                })
+            assertFalse(token.isExpired(),
+                    "Refreshable AccessToken should not expire when expiry is set to 5 minutes later");
+            assertEquals(tokenStr, token.getToken());
+            assertEquals(0, immediateFresher.numCalls());
+        })
                 .verifyComplete();
         tokenCredential.close();
     }
@@ -145,10 +152,53 @@ public class CommunicationTokenCredentialTests {
     }
 
     @Test
+    public void fractionalBackoffAppliedWhenTokenExpiring()
+            throws InterruptedException, ExecutionException, IOException {
+        int validForSecs = 8;
+        double expectedTotalCallsTillLastSecond = Math.floor(Math.log(validForSecs));
+        String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 8);
+        AccessToken accessToken = (new TokenParser()).parseJWTToken(tokenStr);
+        immediateFresher.resetCallCount();
+        immediateFresher.setRefreshedToken(tokenStr);
+        CommunicationTokenRefreshOptions tokenRefreshOptions = new CommunicationTokenRefreshOptions(immediateFresher)
+                .setRefreshProactively(true)
+                .setInitialToken(tokenStr);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
+        long tokenTtlSecs = accessToken.getExpiresAt().toInstant().getEpochSecond() - Instant.now().getEpochSecond();
+        while (tokenTtlSecs > 1) {
+            tokenTtlSecs = accessToken.getExpiresAt().toInstant().getEpochSecond() - Instant.now().getEpochSecond();
+        }
+        tokenCredential.close();
+        assertEquals(expectedTotalCallsTillLastSecond, immediateFresher.numCalls());
+    }
+
+    @Test
+    public void shouldThrowExceptionOnExpiredTokenReturn()
+            throws InterruptedException, ExecutionException, IOException {
+        String expiredToken = tokenMocker.generateRawToken("resourceId", "userIdentity", -5 * 60);
+        immediateFresher.resetCallCount();
+        immediateFresher.setRefreshedToken(expiredToken);
+        CountDownLatch countDownLatch = new CountDownLatch(1);
+        immediateFresher.setOnCallReturn(countDownLatch::countDown);
+        CommunicationTokenRefreshOptions tokenRefreshOptions = new CommunicationTokenRefreshOptions(immediateFresher)
+                .setRefreshProactively(true)
+                .setInitialToken(expiredToken);
+        CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
+
+        countDownLatch.await();
+        assertEquals(1, immediateFresher.numCalls());
+        StepVerifier.create(tokenCredential.getToken())
+                .verifyError(IllegalArgumentException.class);
+        tokenCredential.close();
+    }
+
+    @Test
     public void refresherShouldBeCalledAgainAfterFirstRefreshCall()
             throws InterruptedException, ExecutionException, IOException {
         String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", DEFAULT_EXPIRING_OFFSET_SECONDS);
         immediateFresher.resetCallCount();
+        String newTokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", DEFAULT_EXPIRING_OFFSET_SECONDS + 2);
+        immediateFresher.setRefreshedToken(newTokenStr);
         CountDownLatch firstCountDownLatch = new CountDownLatch(1);
         immediateFresher.setOnCallReturn(firstCountDownLatch::countDown);
         CommunicationTokenRefreshOptions tokenRefreshOptions = new CommunicationTokenRefreshOptions(immediateFresher)
@@ -227,15 +277,20 @@ public class CommunicationTokenCredentialTests {
     }
 
     @Test
-    public void shouldCallbackOnDemandWithoutRefresher() throws InterruptedException, ExecutionException, IOException {
+    public void refresherShouldBeCalledImmediatelyWithoutInitialToken() throws InterruptedException, ExecutionException, IOException {
+        CountDownLatch countDownLatch = new CountDownLatch(1);
+        immediateFresher.setOnCallReturn(countDownLatch::countDown);
         immediateFresher.resetCallCount();
+        String tokenStr = tokenMocker.generateRawToken("resourceId", "userIdentity", 15 * 60);
+        immediateFresher.setRefreshedToken(tokenStr);
         CommunicationTokenRefreshOptions tokenRefreshOptions = new CommunicationTokenRefreshOptions(immediateFresher)
                 .setRefreshProactively(true);
         CommunicationTokenCredential tokenCredential = new CommunicationTokenCredential(tokenRefreshOptions);
+        countDownLatch.await();
+        assertEquals(1, immediateFresher.numCalls());
         StepVerifier.create(tokenCredential.getToken())
                 .assertNext(token -> {
-                    assertEquals(2, immediateFresher.numCalls());
-                    assertFalse(token.isExpired(), "On demand fetching case, should be still valid");
+                    assertFalse(token.isExpired(), "Immediate fetching case, should be still valid");
                 })
                 .verifyComplete();
         tokenCredential.close();