+++ sdk/communication/azure-communication-common/CHANGELOG.md
@@ -1,13 +1,15 @@
 # Release History
 
-## 1.1.0 (Unreleased)
+## 1.1.0-beta.1 (Unreleased)
 
 ### Features Added
 
 - Added new constructor with required param `tokenRefresher` for `CommunicationTokenRefreshOptions`
 - Deprecated old constructor overloads in `CommunicationTokenRefreshOptions` and replaced by fluent setters
 - Added fluent setters for optional properties:
-  - Added `setInitialToken(String initialToken)` setter that allows setting the optional serialized JWT token
+    - Added `setRefreshProactively(boolean refreshProactively)` setter that allows setting whether the token should be proactively renewed prior to its expiry or on demand.
+    - Added `setInitialToken(String initialToken)` setter that allows setting the optional serialized JWT token
+- Added a synchronous token refresher getter `getTokenRefresherSync` for `CommunicationTokenRefreshOptions`
 - Optimization added: When the proactive refreshing is enabled and the token refresher fails to provide a token that's not about to expire soon, the subsequent refresh attempts will be scheduled for when the token reaches half of its remaining lifetime until a token with long enough validity (>10 minutes) is obtained.
 
 ### Breaking Changes